fix: resolve homepage hydration error and restore corrupted auth server logic
