--- packages/shared/src/lib/lib/i18n/config.ts
@@ -1,4 +1,0 @@
-export type Locale = 'en' | 'pl' | 'es' | 'de'
-
-export const locales: Locale[] = ['en', 'pl', 'es', 'de']
-export const defaultLocale: Locale = 'en'
